Rework darken function
Rework darken function in order to make it compatible with chooseColor function.
The issue is coming from the opacity styling when transitioning from the Darken
button to the Color button and vice versa. When you click on one after previously
clicking on the other, the opacity settings to not change along with the button.
This causes incompatibly when going back and forth between the two. The chooseColor
function will also need to be reworked then tested.