Refactor resetSquares function
The function, resetSquares, worked as long as the user entered a number between
1 and 100. If the user entered a number outside of this range, it would alert the user
but not recall the prompt. Refactor the function to include a while loop that will both
alert the user and ensure that the prompt is recalled if the user is to
enter a number outside of the required range.